requirements.txt
python-dotenv==1.0.1
streamlit==1.39.0
pydantic==2.9.2
requests==2.32.3
pandas==2.2.3
streamlit